Regenerate all configure scripts and update the dependent files
Add bakefile and wxCode files needed for updating the build system: copy *.m4
files from Bakefile 0.2.9 (the latest and last 0.x release) and the relevant
contents of wxCode/build into this repository.

Update the paths in build*/configure.ac and Bakefiles.bkgen to refer to the
files in this repository.

Regenerated build*/aclocal.m4 files using the latest aclocal (this, in
particular, means that they just include the files they use instead of
incorporating them) and wxwin.m4.

The latter is especially important as it ensures that wxPdfDocument can be
used with wxGTK3 which wasn't recognized by the macros in the old version of
this file.