New simple autotools-based Unix make system
The new configure and Makefile.am replace the old bakefile-generated makefiles
and wxCode-based configure script.

They are Unix-specific for now, as MSVS projects are used under MSW, but could
be made to work under MSW (or when cross-compiling) in the future.

This commit includes only the files specific to this project, the other files
needed for automake+autoconf to work can be created in the working tree by
running "autoreconf --install" (notice that if wxwin.m4 is not available in
a standard location such as /usr/share/aclocal, you also need to use "-I $wx"
option to let aclocal find it and avoid confusing errors).