Only export selected symbols from the library under Unix
Use ELF visibility to not export everything from the library by default, but
only the symbols marked with WXDLLIMPEXP_PDFDOC. This should work, i.e. export
all the really required symbols, as it already does under MSW and it removes
more than a thousand symbols that should not be exported from the library.