Compile resource files under MSW
Compile .rc files as part of the build by defining a custom rule using
windres.

Notice that Automake makes it quite painful to have a source (.cpp) and
resource (.rc) file with the same name, although this is common practice in
the MSW world, so we have to use a custom extension (.res_o) to avoid object
file name conflicts and hence add these object files manually to CLEANFILES
and LDADD. Renaming the .rc files to use unique names would have been simpler.