Don't link the static library with wxWidgets libraries
It is never necessary to link static libraries with anything and it was, in
fact, actively harmful instead of just useless, at least when cross-compiling
from Linux to macOS, as doing this resulted in completely bogus linker errors
about "file was built for archive which is not the architecture being linked"
(even though it was) and the library being just ignored, resulting in tons of
undefined symbol errors later.

Simply don't link the static library with anything to fix this (it's a pity
that libtool is not smart enough to only apply LIBADD to shared libraries).