Solve issue #44
In the wxMAC build wxPdfDocument detects which font system is available by checking for the symbols wxOSX_USE_CORE_TEXT and wxOSX_USE_ATSU_TEXT. Since wxWidgets 3.x these symbols are no longer defined, although the CoreText font system now seems to be the default. For that reason wxPdfDocument now assumes for wxWidgets 3.x that CoreText is available and uses related methods to identify fonts which are represented by a wxFont object.