Use the transparent background mode by default
Don't set background mode to "solid" in wxPdfDCImpl constructor, this is
inconsistent with the standard wxDC classes, which use transparent
background by default.

Setting it to "solid" went unnoticed for a long time because background
mode was ignored anyhow, but became significant since support for it was
added in 4c54889a6213ab72e660e5a14823da2bf06b1d6c.