Add support for PDF/A-1B conformance (#73)
PDF/A-1B conformance can be enabled (with method wxPdfDocument::SetPdfA1Conformance). However, it is the user's responsibility to use only PDF features allowed for PDF/A-1B conformance. Especially, all used fonts (or font subsets) need to embedded, alpha transparency as well as document protection must not be used.