Remove support for wxWidgets 2.x; replace wxChar type
Almost 10 years after the last release of a wxWidgets 2.x version, namely version 2.8.12, time has come to remove support for it. In this course the use of the character type wxChar was overhauled as well, replacing it by wxUniChar or wxStringCharType as appropriate. The latter should also fix issue #67 hopefully.

Due to this major modification, the next release version will be wxPdfDocument 1.0.0.