Fix wxPdfDC issue with pen and brush color
When using wxPdfDC for creating multi page PDF documents PDF pen and brush were not explicitly set if they happened to be equal to the last pen and brush used on the previous page. This resulted in the use of the default pen color, namely black. On page start PDF pen and brush are now set to an undefined state, forcing them to be synchronized with the DC pen and brush.