Actually use wxPdfDocument::GetImageSize() MIME type parameter
This parameter was not used at all (resulting in a warning) and was
instead shadowed by a local variable (resulting in another warning).

Fix the warnings but, more importantly, actually use the parameter if it
is specified, while still preserving the old behaviour when the default,
empty, value of this parameter is used.