Remove unused "compressed" flag from wxPdfDocument::PutFonts()
It doesn't seem to make much sense to define this variable and then test
for it when it is always true -- because currently compression is always
used for the fonts data.

Just use decompression filter unconditionally and add a comment
explaining that we don't need to have any checks here, notably not for
"m_compress" which is used in the other functions.