Fix problem with page breaks in XML tables
For a table with cells that span across 2 or more rows it could happen that a page break was added for a row in the middle of such a row span. However, it is desirable to keep a block of rows affected by row spanning cells together on a page.